+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Feature: ERC-20 decimals mock up
 - Feature: Wallet ERC-20 portfolio analysis
 - Feature: Solidity event fetcher
+- Feature: Fee estimator
+- Feature: Fetch Uniswap deployment from on-chain data
 
 # 0.1